% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{aggByClone}
\alias{aggByClone}
\title{aggByClone}
\usage{
aggByClone(dat = NULL, removeFeeder = T, ag.fun = sum)
}
\arguments{
\item{dat}{The expression data. If NULL (default), transcript-level count data are used.}

\item{removeFeeder}{Logical; whether to exclude samples from lines grown on feeder cells (default TRUE).}

\item{ag.fun}{Aggregation function; default `sum`}
}
\value{
A list with two slots: `dat` containing the aggregated data, and `annotation` containing the corresponding metadata.
}
\description{
Aggregate expression data from technical replicates
}

