% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{auc}
\alias{auc}
\title{auc}
\usage{
auc(x, y, dens = 100)
}
\arguments{
\item{x}{x values.}

\item{y}{corresponding y values.}

\item{dens}{Number of points.}
}
\value{
The area under the curve.
}
\description{
Computes the area under a curve.
}
