% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{cellpheno.permutateIndividuals}
\alias{cellpheno.permutateIndividuals}
\title{cellpheno.permutateIndividuals}
\usage{
cellpheno.permutateIndividuals(nbIndividuals = 2, nbClone = 2,
  maxTests = 300, logT = FALSE, testfunc = .ttp, agTechRep = mean,
  doSave = TRUE, returnResults = FALSE, ncores = NULL)
}
\arguments{
\item{nbIndividuals}{A positive integer indicating the number of individuals in each group (default 2).}

\item{nbClone}{The number of iPSC clones to consider from each individual (default 2)}

\item{maxTests}{The maximum number of tests to run (default 300).}

\item{logT}{Whether to log-transform the data before testing (default FALSE)}

\item{testfunc}{Function used for testing. Defaults to `.ttp` (t-test)}

\item{agTechRep}{Function used to aggregate technical replicates (default: mean)}

\item{doSave}{Whether to save the results (default TRUE).}

\item{returnResults}{Whether to return the results (default FALSE).}

\item{ncores}{Integer; number of cores to use. Defaults to detected number of cores minus one.}
}
\value{
Nothing (results save to files), or a data.frame if returnResults=TRUE.
}
\description{
Tests for differences in iPSC cellular morphology between groups of random individuals.
}

