% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceExplained.R
\name{cellphenoVarianceExplained}
\alias{cellphenoVarianceExplained}
\title{cellphenoVarianceExplained}
\usage{
cellphenoVarianceExplained(logTransform = TRUE, doPlot = TRUE)
}
\arguments{
\item{logTransform}{Logical; whether to log-transform the data (default TRUE).}

\item{doPlot}{Logical; whether to plot a summary of the results (default TRUE; requires package `pheatmap` to be installed)}
}
\value{
a dataframe containing the square residuals, with cellular features as rows and explanatory variables as columns
}
\description{
Computes for each cellular feature the variance explained by each variable.
}
