% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{combinePermResults}
\alias{combinePermResults}
\title{combinePermResults}
\usage{
combinePermResults(list1, list2, balance = TRUE)
}
\arguments{
\item{list1}{The first list of permutation results.}

\item{list2}{The second list of permutation results.}

\item{balance}{Logical; whether to subsample the largest of the two elements being merged so that the merge is balanced (default TRUE).}
}
\value{
A list of merged permutation results.
}
\description{
Merges the respective elements of two lists of permutation analyses (i.e. output of readPermResults).
}

