% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{DEA.permutateClones}
\alias{DEA.permutateClones}
\title{DEA.permutateClones}
\usage{
DEA.permutateClones(
  nbIndividuals = 2,
  maxTests = 50,
  res = NULL,
  seed = 1,
  addDE = F,
  addDE.foldchanges = c(1.25, 1.5, 2, 3, 5),
  addDE.nbPerFC = 10,
  paired = F,
  doSave = T,
  returnResults = F,
  ncores = NULL,
  quiet = F,
  filter = NULL,
  useExactTest = T,
  DEAfunc = edgeRwrapper
)
}
\arguments{
\item{nbIndividuals}{A positive integer indicating the number of individuals to consider (default 2).}

\item{maxTests}{The maximum number of tests to run (default 50).}

\item{res}{The result of `aggByClone(getGeneExpr())`, optional and used only to speed up multiple calls or use a custom dataset.}

\item{seed}{The seed (for randomization)}

\item{addDE}{Logical; whether to add random differentially-expressed genes (default FALSE).}

\item{addDE.foldchanges}{A numeric vector of foldchanges to add (will use the given foldchanges and their inverse). Default c(1.1,1.25,1.5,2,3,5).}

\item{addDE.nbPerFC}{The number of DEGs for each foldchange (default 5).}

\item{paired}{Logical; whether to run the analysis in a paired way, i.e. `~individual+group` (default FALSE).}

\item{doSave}{Whether to save the results (default TRUE).}

\item{returnResults}{Whether to return the results (default FALSE).}

\item{ncores}{Integer; number of cores to use. Defaults to detected number of cores minus one.}

\item{quiet}{Logical; if TRUE, suppresses progress report (default FALSE)}

\item{filter}{Function; the filter for genes to be tested, to be applied to all rows of the expression matrix (default: no filter). An example value for the filter argument would be `sum(x>10)>2', which would only include genes that have more than 10 counts in more than 2 samples.}

\item{useExactTest}{Logical; whether to use edgeR's exact test for unpaired analysis. Works only with DEAfunc=edgeRwrapper.}

\item{DEAfunc}{Function; the function used to run the differential expression analysis (by default, the `edgeRwrapper' function of this package).}
}
\value{
Nothing (results save to files), or a data.frame if returnResults=TRUE.
}
\description{
Performs differential-expression analyses between groups constituted by corresponding iPSC clones from random individuals (for each individual, one clone is randomly assigned to one group, and the other to the second group).
}
