% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{DEA.permutateIndividuals}
\alias{DEA.permutateIndividuals}
\title{DEA.permutateIndividuals}
\usage{
DEA.permutateIndividuals(
  nbIndividuals = 2,
  maxTests = 50,
  nbClone = 2,
  res = NULL,
  useOnlyMulticlone = TRUE,
  seed = 1,
  addDE = FALSE,
  addDE.foldchanges = c(1.25, 1.5, 2, 3, 5),
  addDE.nbPerFC = 10,
  doSave = TRUE,
  returnResults = !doSave,
  ncores = NULL,
  quiet = FALSE,
  filter = function(x) sum(x > 10) > 2,
  nested = FALSE,
  fasterCombinations = TRUE,
  DEAfunc = edgeRwrapper
)
}
\arguments{
\item{nbIndividuals}{Either a positive integer, or a vector of positive integers of length 2, indicating the number of individuals in each group (default 2).}

\item{maxTests}{The maximum number of tests to run (default 50).}

\item{nbClone}{The number of iPSC clones to consider from each individual (default 2)}

\item{res}{The dataset to use. Should be a list with two slots (`annotation` and `dat`), as the object available in `data("GSE79636")`.
If left NULL, the HipSci dataset is automatically fetched and aggregated (to avoid doing this everything when calling the function with different designs, 
first run `aggByClone(getGeneExpr())` and pass its result via this argument.}

\item{useOnlyMulticlone}{Logical; whether to use only the samples that have more than one clone (default TRUE). Disabling this greatly increase the number of possible combinations, but will make the search for viable combinations _very_ slow. It should be used only if insufficient viable combinations are available (e.g. for nbIndividuals > 5).}

\item{seed}{The seed (for randomization)}

\item{addDE}{Logical; whether to add random differentially-expressed genes (default FALSE).}

\item{addDE.foldchanges}{A numeric vector of foldchanges to add (will use the given foldchanges and their inverse, applied on the first group).}

\item{addDE.nbPerFC}{The number of DEGs for each foldchange.}

\item{doSave}{Whether to save the results (default TRUE).}

\item{returnResults}{Whether to return the results (default FALSE unless `doSave=FALSE`).}

\item{ncores}{Integer; number of cores to use. Defaults to detected number of cores minus one.}

\item{quiet}{Logical; if TRUE, suppresses progress report (default FALSE)}

\item{filter}{Function; the filter for genes to be tested, to be applied to all rows of the expression matrix (default: no filter). An example value for the filter argument would be `sum(x>10)>2', which would only include genes that have more than 10 counts in more than 2 samples.}

\item{nested}{Logical; whether to run the analysis using duplicateCorrelation (default FALSE). Requires `nbClone=2` and `DEAfunc=voomWrapper`.}

\item{fasterCombinations}{Logical; whether to find combinations rapidly by selecting a pre-defined sex ratio (default FALSE). Useful for large datasets.}

\item{DEAfunc}{Function; the function used to run the differential expression analysis (by default, the `edgeRwrapper' function of this package).}
}
\value{
Nothing (results save to files), or a data.frame if returnResults=TRUE.
}
\description{
Performs differential-expression analyses between samples from random individuals.
}
