% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dreamWrapper}
\alias{dreamWrapper}
\title{dreamWrapper}
\usage{
dreamWrapper(e, mm = NULL, nested = NULL, DEsamples = NULL, paired = NULL)
}
\arguments{
\item{e}{the expression matrix}

\item{mm}{Ignored; for consistency with edgeRwrapper.}

\item{nested}{A character vector of size=ncol(e), indicating the individuals.}

\item{DEsamples}{the index of the samples in which the differential expression was added.}

\item{paired}{Ignored; for consistency with edgeRwrapper.}
}
\value{
A data.frame.
}
\description{
Performs one differential expression analysis using the dream method from the
`variancePatition` package (required).
}
