% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{edgeRwrapper}
\alias{edgeRwrapper}
\title{edgeRwrapper}
\usage{
edgeRwrapper(e, mm = NULL, DEsamples = NULL, paired = FALSE, nested = NULL)
}
\arguments{
\item{e}{the expression matrix}

\item{mm}{the design model (as created by the `model.matrix()' function), for paired analyses only. Default NULL.}

\item{DEsamples}{the index of the samples in which the differential expression was added. Mandatory if mm=NULL, ignored otherwise.}

\item{paired}{Logical; whether the analysis should be paired or not.}

\item{nested}{Ignored; for consistency with voomWrapper.}
}
\value{
A data.frame.
}
\description{
Performs one differential expression analysis using edgeR.
}
