% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{fpkm}
\alias{fpkm}
\title{Convert counts to FPKM values}
\usage{
fpkm(counts, mean.frag.size = 500, removeShorter = F)
}
\arguments{
\item{counts}{A numeric dataframe or matrix, with transcripts as rows.}

\item{mean.frag.size}{A numeric value indicating the mean fragment size (default 220) used for effective length calculation.}

\item{removeShorter}{Logical; whether to remove transcripts whose length is shorter than the mean fragment size (default FALSE).}
}
\value{
A dataframe containing the corresponding FPKM values.
}
\description{
Convert fragment counts to FPKM values, using the sum as library size. By default, effective length is used; set mean.frag.size to 0 to use total (given) length.
}

