% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{fpkm2tpm}
\alias{fpkm2tpm}
\title{Convert FPKM values to TPM values}
\usage{
fpkm2tpm(fpkm)
}
\arguments{
\item{fpkm}{A dataframe, matrix or numeric vector of fpkm values.}
}
\value{
The corresponding TPM values.
}
\description{
Convert FPKM values to transcripts per million (TPM).
}

