% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getGeneExpr}
\alias{getGeneExpr}
\title{getGeneExpr}
\usage{
getGeneExpr(unit = "counts", quiet = F)
}
\arguments{
\item{unit}{The expression unit in which to return the data; either "counts" (default), "fpkm", or "tpm".}
}
\description{
Returns the gene-level expression of the HipSci dataset.
}
