% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceExplained.R
\name{getLmeResiduals}
\alias{getLmeResiduals}
\title{getLmeResiduals}
\usage{
getLmeResiduals(x, m, form, variables = NULL)
}
\arguments{
\item{x}{a numeric vector (expression values to be explained)}

\item{m}{a data.frame with rows corresponding to each value of x, and columns indicating the values of explanatory variables.}

\item{form}{Specify the formula to use (required).}

\item{variables}{Character vector containing the names of the variables, in the order in which they should be returned. If NULL, `colnames(m)` is used.}
}
\value{
a numeric vector of squared residuals.
}
\description{
Returns the residuals of x explained by each variable (column of m) using mixed models.
}
