% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{getRecurrentGenes}
\alias{getRecurrentGenes}
\title{getRecurrentGenes}
\usage{
getRecurrentGenes(
  permResults,
  removeInputDE = TRUE,
  varToFetch = "pval.below.05",
  d = NULL
)
}
\arguments{
\item{permResults}{Either the results of the `DEA.permutateIndividuals` or `DEA.permutateClones` function, or a character vector of the files in which the outpout of these functions was save.}

\item{removeInputDE}{Whether to remove the input DEGs (default TRUE)}

\item{varToFetch}{Variable to fetch, any of: `pval.below.05`, `pval.below.01`, `FDR.below.05`, `FDR.below.01` (default `pval.below.05`).}

\item{d}{For internal use.}
}
\value{
A data.frame.
}
\description{
Returns the frequencing of each gene being found differentially expressed in the permutation analyses.
}
