% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getSamplesInfo}
\alias{getSamplesInfo}
\title{getSamplesInfo}
\usage{
getSamplesInfo(removeFeeder = T, doPlot = T)
}
\arguments{
\item{removeFeeder}{Logical; whether to remove the samples cultivated on feeders, which should be exluded from the permutation analysis (Default TRUE).}

\item{doPlot}{Whether to plot summary information on the samples (default TRUE).}
}
\value{
A data.frame.
}
\description{
Returns an annotation of the samples used.
}

