% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensMatrixPlots.R
\name{getSensitivityMatrices}
\alias{getSensitivityMatrices}
\title{getSensitivityMatrices}
\usage{
getSensitivityMatrices(
  reslist,
  bins = 5,
  unlogExpr = TRUE,
  display_numbers = TRUE,
  do.draw = TRUE,
  ...
)
}
\arguments{
\item{reslist}{A list of analysis results, such as produced by the `readPermResults()` function.
The names of the element will be used as heatmap titles. Note that sensitivity can only be plotted 
when the analyses were run using `addDE=TRUE`.}

\item{bins}{Either an integer indicating the number of bins in which to (try to) 
split the expression levels (default 5), or a list of ranges for binning.}

\item{unlogExpr}{Whether to "un-log" read counts for labels (default TRUE).}

\item{do.draw}{Logical; whether to draw the heatmap (default TRUE). If FALSE, will return the Heatmap list object
(assuming ComplexHeatmap is being used - otherwise the argument is ignored)}

\item{...}{Any further argument passed to the plotting function.}
}
\value{
a heatmap.
}
\description{
Plot one or more sensitivity matrices across foldchanges and expression levels.
Will use `ComplexHeatmap` if available, otherwise `pheatmap`.
}
