% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensMatrixPlots.R
\name{getSensitivityMatrix}
\alias{getSensitivityMatrix}
\title{getSensitivityMatrix}
\usage{
getSensitivityMatrix(res, bins = 5, unlogExpr = T, doPlot = T)
}
\arguments{
\item{res}{A list such as one of the items returned by the `readPermResults()` function.}

\item{bins}{Either an integer indicating the number of bins in which to (try to) split the expression levels (default 5), 
or a vectors of breaks defining the bins.}

\item{unlogExpr}{Whether to "un-log" read counts for labels (default TRUE).}

\item{doPlot}{Whether to plot rather than return the matrix (default TRUE).}
}
\value{
either a heatmap or a matrix.
}
\description{
Retrieves and eventually plot a sensitivity matrix across foldchanges and expression levels.
Use the higher-level function `getSensitivityMatrices` instead.
}
