% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{getSpuriousFCs}
\alias{getSpuriousFCs}
\title{getSpuriousFCs}
\usage{
getSpuriousFCs(permResults, pthreshold = 0.05)
}
\arguments{
\item{permResults}{Either the results of the `DEA.permutateIndividuals` or `DEA.permutateClones` function, or a character vector of the files in which the output of these functions was saved.}

\item{pthreshold}{The p-value threshold above which to consider a gene differentially-expressed.}
}
\value{
A vector of log2 foldchanges.
}
\description{
Returns the log2(foldchange) of all spurious DEGs in results of permutations analyses.
}
