% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getTxExpr}
\alias{getTxExpr}
\title{getTxExpr}
\usage{
getTxExpr(unit = "counts")
}
\arguments{
\item{unit}{The expression unit in which to return the data; either "counts" (default), "fpkm", or "tpm".}
}
\description{
Returns the transcript-level expression of the HipSci dataset.
}
