% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{multiProbSpurious}
\alias{multiProbSpurious}
\title{multiProbSpurious}
\usage{
multiProbSpurious(d, FCfield = "logFC", FDRfield = NULL)
}
\arguments{
\item{d}{A data.frame of differential expression results, but HGNC symbols as row.names and a column indicating the log2 foldchange.}

\item{FCfield}{The name of the column in `d` containing the log2(foldchange) (default `logFC`).}

\item{FDRfield}{Optional; the name of the column in `d` containing the FDR. If given, a column will be added with the maximum of the FDR and the probability of DE in the permuations. (default disabled)}
}
\value{
The data.frame d, with the addition columns probDEinPermutations and probFCinPermutations
}
\description{
Annotates a data.frame of DEA results (such as produced by edgeR's `topTags` function) with the probability of each gene and foldchange in the DEA permutations.
}

