% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{probSpurious}
\alias{probSpurious}
\title{probSpurious}
\usage{
probSpurious(geneSymbol = NULL, log2FC = NULL)
}
\arguments{
\item{geneSymbol}{A HGNC gene symbol}

\item{log2FC}{The log2(foldchange) in your test (optional, but either geneSymbol or log2FC must be given)}
}
\value{
The probability of the gene and/or foldchange in the permutation analyses.
}
\description{
Returns the probability of a given gene and/or foldchange in the permutation analyses.
}

