% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{readPermResults}
\alias{readPermResults}
\title{readPermResults}
\usage{
readPermResults(resFiles)
}
\arguments{
\item{resFiles}{character vector indicating the names of the results files (.RData files save by the DEA permutation functions)}
}
\value{
A list with the following slots:
`nbComps: number of comparisons.
`FP: number of false (or spurious) positives for each comparison.
`TP: (only if the permutations were generated with addDE=TRUE) The number of true positives for each comparison.
`DEGs: (only if the permutations were generated with addDE=TRUE) a data.frame containing, for each gene, the number of times it was found with a FDR below 0.05, the absLog2FC introduced in the input, and the mean fragment count.
}
\description{
readPermResults
}

