% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{TMM}
\alias{TMM}
\title{TMM}
\usage{
TMM(dat)
}
\arguments{
\item{dat}{The data to be normalized.}
}
\value{
A dataframe of normalized read counts.
}
\description{
A wrapper for edgeR's TMM normalization
}

