% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceExplained.R
\name{transcriptionalVarianceExplained}
\alias{transcriptionalVarianceExplained}
\title{transcriptionalVarianceExplained}
\usage{
transcriptionalVarianceExplained(res = NULL, unit = "tpm",
  logTransform = T, variables = c("sex", "individual"), doPlot = TRUE,
  mixed = TRUE, formula = NULL, ncores = NULL)
}
\arguments{
\item{unit}{Either "tpm" (default), "fpkm", or "counts".}

\item{logTransform}{Logical; whether to log-transform the data (default T).}

\item{variables}{The explanatory variables to include. Default "sex" and "individual"}

\item{doPlot}{Logical; whether to plot a summary of the results (default TRUE; requires packages `beanplot` and `LSD` to be installed)}

\item{mixed}{Logical; whether to use mixed models (via `lme4`) (default TRUE).}

\item{formula}{Specify the formula to use. If specified, this overrides the `variables` argument. If `mixed=F`, the default formula is `~.`; if `mixed=T`, each variable is considered a random variable.}

\item{ncores}{The number of threads (by defaults uses all but 1 cores)}
}
\value{
a dataframe containing the square residuals, with genes as rows and explanatory variables as columns (plus a first column containing the gene's mean expression).
}
\description{
Computes for each gene the variance explained by each variable.
}

