% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{vstLmerWrapper}
\alias{vstLmerWrapper}
\title{vstLmerWrapper}
\usage{
vstLmerWrapper(e, nested = NULL, DEsamples = NULL, mm = NULL, paired = NULL)
}
\arguments{
\item{e}{the expression matrix}

\item{nested}{NULL or character vector of size=ncol(e), indicating the blocking variable (default NULL).}

\item{DEsamples}{the index of the samples in which the differential expression was added. Mandatory if mm=NULL, ignored otherwise.}

\item{mm}{Ignored; for consistency with alternative functions.}

\item{paired}{Ignored; for consistency with alternative functions.}
}
\value{
A data.frame.
}
\description{
Performs one differential expression analysis using mixed models (via `lme4::lmer`) on `DESeq2` vst-transformed counts
}
