% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.permutations.R
\name{DEA.permutateIndividuals}
\alias{DEA.permutateIndividuals}
\title{DEA.permutateIndividuals}
\usage{
DEA.permutateIndividuals(nbIndividuals = 2, maxTests = 300, nbClone = 2,
  res = NULL, useOnlyMulticlone = TRUE, seed = 1, addDE = F,
  addDE.foldchanges = c(1.25, 1.5, 2, 3, 5), addDE.nbPerFC = 10,
  doSave = T, returnResults = F, ncores = NULL, quiet = F,
  filter = NULL, nested = FALSE, DEAfunc = edgeRwrapper)
}
\arguments{
\item{nbIndividuals}{Either a positive integer, or a vector of positive integers of length 2, indicating the number of individuals in each group (default 2).}

\item{maxTests}{The maximum number of tests to run (default 300).}

\item{nbClone}{The number of iPSC clones to consider from each individual (default 2)}

\item{res}{The result of `aggByClone(getGeneExpr())`, optional and used only to speed up multiple calls or use a custom dataset.}

\item{useOnlyMulticlone}{Logical; whether to use only the samples that have more than one clone (default TRUE). Disabling this greatly increase the number of possible combinations, but will make the search for viable combinations _very_ slow. It should be used only if insufficient viable combinations are available (e.g. for nbIndividuals > 5).}

\item{seed}{The seed (for randomization)}

\item{addDE}{Logical; whether to add random differentially-expressed genes (default FALSE).}

\item{addDE.foldchanges}{A numeric vector of foldchanges to add (will use the given foldchanges and their inverse, applied on the first group). Default c(1.1,1.25,1.5,2,3,5).}

\item{addDE.nbPerFC}{The number of DEGs for each foldchange (default 5).}

\item{doSave}{Whether to save the results (default TRUE).}

\item{returnResults}{Whether to return the results (default FALSE).}

\item{ncores}{Integer; number of cores to use. Defaults to detected number of cores minus one.}

\item{quiet}{Logical; if TRUE, suppresses progress report (default FALSE)}

\item{filter}{Function; the filter for genes to be tested, to be applied to all rows of the expression matrix (default: no filter). An example value for the filter argument would be `sum(x>10)>2', which would only include genes that have more than 10 counts in more than 2 samples.}

\item{nested}{Logical; whether to run the analysis using duplicateCorrelation (default FALSE). Requires `nbClone=2` and `DEAfunc=voomWrapper`.}

\item{DEAfunc}{Function; the function used to run the differential expression analysis (by default, the `edgeRwrapper' function of this package).}
}
\value{
Nothing (results save to files), or a data.frame if returnResults=TRUE.
}
\description{
Performs differential-expression analyses between samples from random individuals.
}

