% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceExplained.R
\name{getAOVresiduals}
\alias{getAOVresiduals}
\title{getAOVresiduals}
\usage{
getAOVresiduals(x, m)
}
\arguments{
\item{x}{a numeric vector (expression values to be explained)}

\item{m}{a data.frame with rows corresponding to each value of x, and columns indicating the values of explanatory variables.}
}
\value{
a numeric vector of squared residuals.
}
\description{
Returns the residuals of x explained by each variable (column of m).
}

