% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceExplained.R
\name{transcriptionalVarianceExplained}
\alias{transcriptionalVarianceExplained}
\title{transcriptionalVarianceExplained}
\usage{
transcriptionalVarianceExplained(unit = "tpm", logTransform = T,
  variables = c("sex", "individual", "clone", "replicate"), doPlot = TRUE,
  ncores = NULL)
}
\arguments{
\item{unit}{Either "tpm" (default), "fpkm", or "counts".}

\item{logTransform}{Logical; whether to log-transform the data (default T).}

\item{variables}{The explanatory variables to include. Should be any combination of "sex", "individual", "clone", and "replicate".}

\item{doPlot}{Logical; whether to plot a summary of the results (default TRUE; requires packages `beanplot` and `LSD` to be installed)}

\item{ncores}{The number of threads (by defaults uses all but 1 cores)}
}
\value{
a dataframe containing the square residuals, with genes as rows and explanatory variables as columns (plus a first column containing the gene's mean expression).
}
\description{
Computes for each gene the variance explained by each variable.
}

